Refactoring to separate server and db code
