+++ Dockerfile
@@ -8,6 +8,8 @@ COPY package-lock.json .
 RUN npm install
 
 COPY index.js .
+COPY db.js .
+COPY server.js .
 
 EXPOSE 3000
