+++ db.js
@@ -1,0 +1,9 @@
+var DataStore = require('nedb');
+var DB_FILE_NAME = __dirname + "/contacts.json";
+
+var db = new DataStore({
+    filename: DB_FILE_NAME,
+    autoload: true
+});
+
+module.exports = db;
